Move #6: Modifier의 확장함수 dropShadow를 Modifier.kt 파일로 이동
